requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.8.0
redis==5.0.7
sqlparse==0.5.0
